Add sort-type module
Move SortType enum and related functions to sort-type.
