+++ src/meson.build
@@ -153,6 +153,8 @@ main_sources = files('advanced-exif.cc',
 'similar.h',
 'slideshow.cc',
 'slideshow.h',
+'sort-type.cc',
+'sort-type.h',
 'thumb.cc',
 'thumb.h',
 'thumb-standard.cc',
