+++ src/pan-view/pan-calendar.cc
@@ -36,7 +36,6 @@
 #include "pan-util.h"
 #include "pan-view-filter.h"
 #include "pan-view.h"
-#include "typedefs.h"
 
 namespace
 {
