+++ src/pan-view/pan-folder.cc
@@ -31,7 +31,6 @@
 #include "pan-types.h"
 #include "pan-util.h"
 #include "pan-view-filter.h"
-#include "typedefs.h"
 
 static void pan_flower_size(PanWindow *pw, gint &width, gint &height)
 {
