+++ src/pan-view/pan-grid.cc
@@ -27,7 +27,6 @@
 #include "pan-types.h"
 #include "pan-util.h"
 #include "pan-view-filter.h"
-#include "typedefs.h"
 
 void pan_grid_compute(PanWindow *pw, FileData *dir_fd, gint &width, gint &height)
 {
