+++ src/sort-type.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2004 John Ellis
+ * Copyright (C) 2006 John Ellis
  * Copyright (C) 2008 - 2016 The Geeqie Team
  *
  * Author: John Ellis
@@ -19,27 +19,27 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-#ifndef MENU_H
-#define MENU_H
+#ifndef SORT_TYPE_H
+#define SORT_TYPE_H
 
-#include <glib-object.h>
 #include <glib.h>
-#include <gtk/gtk.h>
 
-enum SortType : gint;
+enum SortType : gint {
+	SORT_NONE,
+	SORT_NAME,
+	SORT_SIZE,
+	SORT_TIME,
+	SORT_CTIME,
+	SORT_PATH,
+	SORT_NUMBER,
+	SORT_EXIFTIME,
+	SORT_EXIFTIMEDIGITIZED,
+	SORT_RATING,
+	SORT_CLASS
+};
 
-gpointer submenu_item_get_data(GtkWidget *submenu_item);
+gchar *sort_type_get_text(SortType method);
+bool sort_type_requires_metadata(SortType method);
 
-GtkWidget *submenu_add_edit(GtkWidget *menu, gboolean sensitive, GList *fd_list, GCallback func, gpointer data);
-
-GtkWidget *submenu_add_sort(GtkWidget *menu, GCallback func, gpointer data,
-                            gboolean show_current, SortType type);
-
-GtkWidget *submenu_add_alter(GtkWidget *menu, GCallback func, gpointer data);
-
-GtkWidget *submenu_add_collections(GtkWidget *menu, gboolean sensitive,
-                                   GCallback func, gpointer data);
-
-void popup_menu_bar(GtkWidget *widget, GCallback expander_height_cb);
 #endif
 /* vim: set shiftwidth=8 softtabstop=0 cindent cinoptions={1s: */
